Python/Pyrogram/TheOCRBot/requirements.txt
pyrogram
python-dotenv
tgcrypto
motor
aiohttp
